Config Security in Endpoints

+++ src/main/java/com/lautadev/juegos_deportivos/service/AccountService.java
@@ -1,10 +1,13 @@
 package com.lautadev.juegos_deportivos.service;
 
 import com.lautadev.juegos_deportivos.model.Account;
+import com.lautadev.juegos_deportivos.model.Enroller;
 import com.lautadev.juegos_deportivos.model.Role;
 import com.lautadev.juegos_deportivos.repository.IAccountRepository;
+import com.lautadev.juegos_deportivos.repository.IEnrollerRepository;
 import com.lautadev.juegos_deportivos.util.NullAwareBeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -20,6 +23,12 @@ public class AccountService implements IAccountService{
 
     @Autowired
     private IRoleService roleService;
+
+    @Autowired
+    private IEnrollerRepository enrollerRepository;
+
+    @Autowired
+    private IUserDetailsService userDetailsService;
 
     @Override
     public Account saveAccount(Account account) {
@@ -54,16 +63,27 @@ public class AccountService implements IAccountService{
 
     @Override
     public void deleteAccount(Long id) {
+        Account account = this.findAccount(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        Enroller enroller = enrollerRepository.findById(enrollerId).orElse(null);
+        if(!account.getId().equals(enroller.getAccount().getId())){
+            throw new AccessDeniedException("You are not authorized to delete this account");
+        }
         accountRepository.deleteById(id);
     }
 
     @Override
     public Account editAccount(Long id,Account account) {
         Account accountEdit = this.findAccount(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        Enroller enroller = enrollerRepository.findById(enrollerId).orElse(null);
+        if(!accountEdit.getId().equals(enroller.getAccount().getId())){
+            System.out.println("Acceso denegado papu :v");
+            throw new AccessDeniedException("You are not authorized to edit this account");
+        }
 
         NullAwareBeanUtils.copyNonNullProperties(account,accountEdit);
 
-        assert accountEdit != null;
         return this.saveAccount(accountEdit);
     }
