+++ src/main/java/com/lautadev/juegos_deportivos/service/EnrollerService.java
@@ -5,6 +5,7 @@ import com.lautadev.juegos_deportivos.model.Enroller;
 import com.lautadev.juegos_deportivos.repository.IEnrollerRepository;
 import com.lautadev.juegos_deportivos.util.NullAwareBeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -15,6 +16,9 @@ import java.util.Optional;
 public class EnrollerService implements IEnrollerService {
     @Autowired
     private IEnrollerRepository enrollerRepository;
+
+    @Autowired
+    private IUserDetailsService userDetailsService;
 
     @Override
     public void saveEnroller(Enroller enroller) {
@@ -41,12 +45,22 @@ public class EnrollerService implements IEnrollerService {
 
     @Override
     public void deleteEnroller(Long id) {
+        Enroller enroller = enrollerRepository.findById(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        if(!enroller.getId().equals(enrollerId)){
+            throw new AccessDeniedException("You are not authorized to delete this enroller");
+        }
         enrollerRepository.deleteById(id);
     }
 
     @Override
     public Optional<EnrollerDTO> editEnroller(Long id, Enroller enroller) {
         Enroller enrollerEdit = enrollerRepository.findById(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        if(!enrollerEdit.getId().equals(enrollerId)){
+            System.out.println("Acceso denegado papu :v");
+            throw new AccessDeniedException("You are not authorized to edit this enroller");
+        }
 
         NullAwareBeanUtils.copyNonNullProperties(enroller,enrollerEdit);
