+++ src/main/java/com/lautadev/juegos_deportivos/service/InscriptionService.java
@@ -6,6 +6,7 @@ import com.lautadev.juegos_deportivos.repository.IInscriptionRepository;
 import com.lautadev.juegos_deportivos.util.NullAwareBeanUtils;
 import com.lautadev.juegos_deportivos.util.PDFGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -16,6 +17,9 @@ import java.util.Optional;
 public class InscriptionService implements IInscriptionService {
     @Autowired
     private IInscriptionRepository inscriptionRepository;
+
+    @Autowired
+    private IUserDetailsService userDetailsService;
 
     @Override
     public void saveInscription(Inscription inscription) {
@@ -42,16 +46,24 @@ public class InscriptionService implements IInscriptionService {
 
     @Override
     public void deleteInscription(Long id) {
+        Inscription inscription = inscriptionRepository.findById(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        if (!inscription.getEnroller().getId().equals(enrollerId)) {
+            throw new AccessDeniedException("You are not authorized to delete this inscription");
+        }
         inscriptionRepository.deleteById(id);
     }
 
     @Override
     public InscriptionDTO editInscription(Long id,Inscription inscription) {
         Inscription inscriptionEdit = this.findInscription(id).orElse(null);
+        Long enrollerId = userDetailsService.getCurrentEnrollerId();
+        if (!inscriptionEdit.getEnroller().getId().equals(enrollerId)) {
+            throw new AccessDeniedException("You are not authorized to edit this inscription");
+        }
 
         NullAwareBeanUtils.copyNonNullProperties(inscription,inscriptionEdit);
 
-        assert inscriptionEdit != null;
         inscriptionRepository.save(inscriptionEdit);
         return this.findInscriptionDTO(inscriptionEdit.getId());
     }
